+++ lib/flutter_starter_template_cli.dart
@@ -1,5 +1,6 @@
 export 'src/config_resolver.dart';
 export 'src/feature_names.dart';
+export 'src/rewrite/firebase.dart';
 export 'src/project_config.dart';
 export 'src/rewrite/clean_slate.dart';
 export 'src/rewrite/feature_template.dart';